// cdam-certify/frontend/src/app/core/models/enums.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  VIEWER = 'VIEWER',
}

export enum ProgramType {
  SHORT_COURSE = 'SHORT_COURSE',
  INTERNSHIP = 'INTERNSHIP',
  ATTACHMENT = 'ATTACHMENT',
}

export enum EnrollmentStatus {
  APPLIED = 'APPLIED',
  ENROLLED = 'ENROLLED',
  COMPLETED = 'COMPLETED',
  WITHDRAWN = 'WITHDRAWN',
}

export enum CertificateStatus {
  VALID = 'VALID',
  REVOKED = 'REVOKED',
}